+++ i386/libsaio/hfs.h
@@ -20,14 +20,16 @@
  * @APPLE_LICENSE_HEADER_END@
  */
 
+#include <sys/types.h>
+
 extern long HFSInitPartition(CICell ih);
 extern long HFSLoadFile(CICell ih, char * filePath);
-extern long HFSReadFile(CICell ih, char * filePath, void *base, uint64_t offset, uint64_t length);
+extern long HFSReadFile(CICell ih, char * filePath, void *base, u_int64_t offset, u_int64_t length);
 extern long HFSGetDirEntry(CICell ih, char * dirPath, long long * dirIndex,
-                           char ** name, long * flags, long * time,
+                           char ** name, long * flags, u_int32_t * time,
                            FinderInfo * finderInfo, long * infoValid);
 extern void HFSGetDescription(CICell ih, char *str, long strMaxLen);
-extern long HFSGetFileBlock(CICell ih, char *str, unsigned long long *firstBlock);
+extern long HFSGetFileBlock(CICell ih, char *str, u_int64_t *firstBlock);
 extern long HFSGetUUID(CICell ih, char *uuidStr);
 extern void HFSFree(CICell ih);
 extern bool HFSProbe (const void *buf);
